move gamestate helpers to new folder
