+++ src/OpenLoco/Input/ShortcutManager.cpp
@@ -1,9 +1,9 @@
 #include "ShortcutManager.h"
 #include "../CompanyManager.h"
 #include "../GameCommands/GameCommands.h"
+#include "../GameStateHelpers/LastGameOption.h"
 #include "../Input.h"
 #include "../Interop/Interop.hpp"
-#include "../LastGameOption.h"
 #include "../Localisation/StringIds.h"
 #include "../S5/S5.h"
 #include "../StationManager.h"
