+++ src/OpenLoco/Map/TileManager.cpp
@@ -4,10 +4,10 @@
 #include "../Entities/Misc.h"
 #include "../Game.h"
 #include "../GameState.h"
+#include "../GameStateHelpers/LastGameOption.h"
 #include "../IndustryManager.h"
 #include "../Input.h"
 #include "../Interop/Interop.hpp"
-#include "../LastGameOption.h"
 #include "../Map/Map.hpp"
 #include "../Objects/BuildingObject.h"
 #include "../TownManager.h"
