+++ src/OpenLoco/Windows/BuildVehicle.cpp
@@ -4,11 +4,11 @@
 #include "../Entities/EntityManager.h"
 #include "../GameCommands/GameCommands.h"
 #include "../GameState.h"
+#include "../GameStateHelpers/LastGameOption.h"
 #include "../Graphics/Colour.h"
 #include "../Graphics/ImageIds.h"
 #include "../Input.h"
 #include "../Interop/Interop.hpp"
-#include "../LastGameOption.h"
 #include "../Localisation/FormatArguments.hpp"
 #include "../Localisation/StringIds.h"
 #include "../Map/Map.hpp"
