+++ src/OpenLoco/Windows/Construction/Common.cpp
@@ -2,10 +2,10 @@
 #include "../../Date.h"
 #include "../../GameCommands/GameCommands.h"
 #include "../../GameState.h"
+#include "../../GameStateHelpers/LastGameOption.h"
 #include "../../Graphics/Colour.h"
 #include "../../Graphics/ImageIds.h"
 #include "../../Input.h"
-#include "../../LastGameOption.h"
 #include "../../Map/Track/TrackData.h"
 #include "../../Objects/AirportObject.h"
 #include "../../Objects/BridgeObject.h"
