+++ src/OpenLoco/Windows/TerraForm.cpp
@@ -3,11 +3,11 @@
 #include "../Economy/Economy.h"
 #include "../GameCommands/GameCommands.h"
 #include "../GameState.h"
+#include "../GameStateHelpers/LastGameOption.h"
 #include "../Graphics/Colour.h"
 #include "../Graphics/ImageIds.h"
 #include "../Input.h"
 #include "../Interop/Interop.hpp"
-#include "../LastGameOption.h"
 #include "../Localisation/FormatArguments.hpp"
 #include "../Localisation/StringIds.h"
 #include "../Map/Map.hpp"
