+++ src/OpenLoco/Windows/ToolbarTop.cpp
@@ -4,12 +4,12 @@
 #include "../Entities/EntityManager.h"
 #include "../Game.h"
 #include "../GameCommands/GameCommands.h"
+#include "../GameStateHelpers/LastGameOption.h"
 #include "../Graphics/Colour.h"
 #include "../Graphics/Gfx.h"
 #include "../Graphics/ImageIds.h"
 #include "../Input.h"
 #include "../Interop/Interop.hpp"
-#include "../LastGameOption.h"
 #include "../Localisation/StringIds.h"
 #include "../MultiPlayer.h"
 #include "../Objects/InterfaceSkinObject.h"
